// projects/projects/src/api/model/learningPath.ts
/**
 * MaestroHost - SceneService
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Scenario } from './scenario';
import { Scene } from './scene';
import { Target } from './target';

export interface LearningPath {
  learningPathId?: number;
  name?: string;
  description?: string;
  scene?: Scene;
  numberOfTargets?: number;
  numberOfVisits?: number;
  createdAtUTC?: Date;
  inUse?: boolean;
  scenarios?: Array<Scenario>;
  targets?: Array<Target>;
  readonly sceneId?: number;
}